Combine user/userconfig models (Nautobot 1.0.0b3), remove invalid circuit cabling loop in test data
