+++ nautobot_netbox_importer/diffsync/models/contenttypes.py
@@ -14,3 +14,9 @@ class ContentType(NautobotBaseModel):
 
     app_label: str
     model: str
+
+    def __init__(self, *args, app_label=None, model=None, **kwargs):
+        """Map NetBox 'auth.user' content type to Nautobot 'users.user' content type."""
+        if app_label == "auth" and model == "user":
+            app_label = "users"
+        super().__init__(*args, app_label=app_label, model=model, **kwargs)
